+++ guillotina/db/transaction_manager.py
@@ -116,14 +116,29 @@ class TransactionManager:
             await self._close_txn(txn)
 
     async def _close_txn(self, txn: typing.Optional[ITransaction]):
-        if txn is not None and txn._db_conn is not None:
-            try:
-                txn._query_count_end = txn.get_query_count()
-            except AttributeError:
-                pass
+        conn = None
+        if txn is not None:
+            # Since we are about to release the connection to the pool, we need to ensure
+            # that there isn't any other task using the same connection during the process.
+            # Otherwise, this would produce the exception:
+            #  InterfaceError: cannot perform operation: another operation is in progress
+            async with txn.lock:
+                try:
+                    txn._query_count_end = txn.get_query_count()
+                except AttributeError:
+                    pass
+
+                if txn._db_txn is not None:  # type: ignore
+                    raise RuntimeError("db txn is not none and we are about to release the db connection")
+
+                conn = txn._db_conn
+                # At this point the connection is no longer reserved by this transaction.
+                txn._db_conn = None
+
+        if conn is not None:
             try:
                 try:
-                    await self._storage.close(txn._db_conn)
+                    await self._storage.close(conn)
                 except asyncpg.exceptions.InterfaceError as ex:
                     if "received invalid connection" in str(ex):
                         # ignore, new pool was created so we can not close this conn
@@ -132,22 +147,20 @@ class TransactionManager:
                         raise
                 except asyncpg.exceptions.InternalClientError:
                     # edge-case where connection is already released
-                    if txn._db_conn is not None:
+                    if conn is not None:
                         raise
             except Exception:
                 # failsafe terminate to make sure connection is cleaned
-                if txn._db_conn is None:
+                if conn._con is None:
                     raise
-                if txn._db_conn._con is None:
-                    raise
+
                 try:
-                    await self._storage.terminate(txn._db_conn)
+                    await self._storage.terminate(conn)
                 except asyncpg.exceptions.InterfaceError as ex:
                     if "released back to the pool" in str(ex):
                         pass
                     else:
                         raise
-            txn._db_conn = None
 
     async def abort(self, *, txn: typing.Optional[ITransaction] = None) -> None:
         try: