+++ guillotina/factory/app.py
@@ -230,7 +230,6 @@ async def startup_app(config_file=None, settings=None, loop=None, server_app=Non
     configure.scan("guillotina.annotations")
     configure.scan("guillotina.constraintypes")
     configure.scan("guillotina.subscribers")
-    configure.scan("guillotina.db.strategies")
     configure.scan("guillotina.db.storages.vacuum")
     configure.scan("guillotina.db.cache")
     configure.scan("guillotina.db.writer")