+++ guillotina/factory/content.py
@@ -172,7 +172,7 @@ class Database:
         txn = await tm.begin()
 
         try:
-            await txn._strategy.retrieve_tid()
+            await txn.initialize_tid()
             root = await tm._storage.load(txn, ROOT_ID)
             if root is not None:
                 root = app_settings["object_reader"](root)