+++ CHANGELOG.rst
@@ -4,6 +4,14 @@ CHANGELOG
 6.4.0rc1 (unreleased)
 ---------------------
 
+- Some fixes and improvements to the new TransactionConnectionContextManager
+  [masipcat]
+
+- Port the following changes:
+  - Remove db transaction strategy support
+  - Try not to reserve connection when possible
+    [vangheem]
+
 - Black to 21.10b0 version
   [bloodbare]